Add Ecto custom type to handle slug as id
This commit add a custom type for Ecto which instruct it how to handle
bynary id containing more than just an integer. So now when a model
uses a "permalink" type primary key. It will be converted from binary
to integer bu just keeping the leading integer in the binary. If the
binary doesn't start with a integer then an error is returned.