Persist annotation on submition
When an annotation is pushed to the video channel it try to persist it
to the database and to assign it to the current user. This persisted
info are then used to display back the newly created annotation to all
channel subscribers and it uses persisted info to fill the annotation.